pyproject.toml
[project]
name = "spotify-album-art-matrix"
version = "0.0.1"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.1",
]
